[Core] feat: Widen Request.priority from int to float
Allow fractional priority values (e.g. 0.5, 1.7) in addition to
integers. This is a backward-compatible type annotation change:
existing callers passing integer priorities (0, 1, -1) see identical
behavior since float comparisons preserve integer ordering.

Motivation: integer priority limits expressiveness to a handful of
discrete levels. Float priority enables finer-grained prioritization
and lays the groundwork for dynamic priority mechanisms such as
aging (see RFC #6077) and composite scoring (see RFC #29406).

Changes:
- vllm/v1/request.py: priority: int -> float
- vllm/v1/engine/__init__.py: EngineCoreRequest.priority: int -> float
- vllm/v1/engine/llm_engine.py: add_request signature
- vllm/v1/engine/async_llm.py: generate/add_request signatures (4)
- vllm/v1/engine/input_processor.py: add_request signature
- tests/v1/core/test_scheduler.py: Sequence import, helper type widened,
  two new tests for float ordering

Note: OpenAI-compatible API protocol files (entrypoints/) retain int
for now to preserve the existing API contract; float support there
is a follow-up.

Duplicate check: no existing open PR addresses this specific change.
Related RFCs (#6077, #29406) discuss broader scheduling improvements
but do not include this type-widening.

+++ vllm/v1/engine/llm_engine.py
@@ -224,7 +224,7 @@ class LLMEngine:
         lora_request: LoRARequest | None = None,
         tokenization_kwargs: dict[str, Any] | None = None,
         trace_headers: Mapping[str, str] | None = None,
-        priority: int = 0,
+        priority: float = 0.0,
         session_id: str | None = None,
         prompt_text: str | None = None,
     ) -> str:
